FEAT: Add static s3 bootstrap file
AWS EMR requires the bootstrap actions to be contained in S3, but this
prevents easy tracking using a VCS.

Create a static file that can be placed in S3, and called with VCS info
to run a dynamic script that is in version control. This means the
version of bootstrap scripts is set in the cloudformation template.