CHORE: Disable deploying to pypi for testing
The next few commits and tags are likely to fail while I work out how to
deploy to Maven. I don't want to be deploying to pypi while this is
being worked out.

Disable the call to deploy to pypi temporarily. Revert this commit once
the rest of the deployment process is working.