CHORE: Reference timestamp for hail-is snapshot
Maven will no longer accept SNAPSHOT versions as dependencies, but the
available version of hail-is is only published as a snapshot. However
the hosted file is referenced with a timestamp, and according to
https://hsgamer.me/posts/snapshot-dependency-maven-central
that can be used as a version.

Use the snapshot timestamp as the version, which will hopefully
(although the path still contains /0.2.74-SNAPSHOT/) allow the correct
jar file to be accessed.

If this doesn't succeed, we're looking at attempting a full publishing
of the hail-is library, which needs to include signing, licenses,
sources, javadocs and other metadata. In this case we're likely to
simply skip 0.5.4 and move straight to attempting to publish the
hail-free version.