FIX: Ignore directories in example/ in setup.py
When creating the data_files argument to setup, listed objects should
only be files. In 1504b8b the examples directory was restructured to
include subdirectories, which are then included and cause setup.py to
crash.

Filter out directory objects when listing files for the data_files
argument to setup in setup.py.