CHORE: Add debugging flags to mvn call
Since running the github action for
ea6268c3f59fb07a2d32fc514a2a0b9fe7548876, the call to mvn deploy now
fails when trying to download the hail-is snapshot pom file. As no maven
related code has changed, and the file appears to still be accessible,
more context is needed.

Add full debug logging  with -X and a stacktrace of the failed call with
-e.