CHORE: Remove timestamp from hail-is version
Maven seems to be (correctly) interpreting the timestamp at the end of
the hail-is version to indicate that the artifact is a snapshot. This
appears to be causing it to be install as "0.2.74-SNAPSHOT" which is not
compatible with the version requirement in the pom.xml, or the maven
repository's requirement to not have SNAPSHOT dependencies.

Remove the timestamp from the hail-is version, to make it appear as a
release version.